Student Record Automation
Today is growing world there is no time to have gat all information from college manually. So we shall have to be changed according to the changing world scenario otherwise we would be outdated. According to the changing we have to accept the new technology and implement these into or day-to-day life. Computers have become very important, but computer cannot work everywhere in the same situation. We need customized application software, which solves our problems very efficiently. So my project is here fully based on the requirement to any college. It can partially solve the problems of college. With the help of this project with the mentioned college can maintain its day-to-day records of the students etc.


TECHNICAL SPECIFICATIONS: 
1. Designing language : HTML,CSS 
2. Developing language: JavaScript 
3. Database language: SQL 
4. Database Tool: MS Access 
5. Operating System: Windows 7/ Windows XP